# docs/openapi/enterprise-agent.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise Agent API",
    "version": "0.1.0",
    "description": "Production-oriented API contract baseline for the enterprise agent platform."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/audit-events": {
      "get": {
        "tags": [
          "Security"
        ],
        "summary": "List auth admin audit events",
        "operationId": "listAuthAuditEvents",
        "description": "Requires permission: auth:manage",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthAuditEventListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/auth/security/anomalies": {
      "get": {
        "tags": [
          "Security"
        ],
        "summary": "List security anomaly events",
        "operationId": "listSecurityAnomalies",
        "description": "Requires permission: auth:manage",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityAnomalyEventListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/auth/security/anomalies/{eventId}/acknowledge": {
      "post": {
        "tags": [
          "Security"
        ],
        "summary": "Acknowledge security anomaly event",
        "operationId": "acknowledgeSecurityAnomaly",
        "description": "Requires permission: auth:manage",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeSecurityAnomalyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityAnomalyEvent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AuthAuditEventListResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "nextOffset": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "limit",
          "offset"
        ]
      },
      "SecurityAnomalyEvent": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "critical"
            ]
          },
          "category": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "acknowledged": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "severity",
          "category",
          "action",
          "message",
          "metadata",
          "acknowledged",
          "createdAt"
        ]
      },
      "SecurityAnomalyEventListResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecurityAnomalyEvent"
            }
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 500
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "nextOffset": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "limit",
          "offset"
        ]
      },
      "AcknowledgeSecurityAnomalyRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "comment": {
            "type": "string"
          }
        }
      }
    }
  }
}